Moved game logic into separate class, added next/prev move feature
* src/game.[h|cc]: This class will become our central game controller. It knows
about the board and the scene view widget. It can also control the chess
position shown on the board.

* src/miniature.[h|cc]: Added a simplistic start menu and modified to use the
new game controller. Also added next move/previous move entries, since it is
not even sure yet whether we want this functionality in the UI. It's easy to
hide/remove when it is in a menu.

* src/board.[h|cc]: The board can now clean itself, and in fact, does so
whenever it is asked to draw a position.

* Makefile.am: Added game sources as a dependency for miniature.