Adapted board rendering to new board model, added logic test for rook
* Makefile.am: Added MLogicAnalyzer dependencies.

* src/board_view.[h|cc], src/graphics_board_item.[h|cc]: Renamed pieceMoved
signal to pieceMoveRequest since MBoardView (more precisely,
MGraphicsBoardItem) is no longer allowed to move pieces on its own. Instead, it
sends a request to the MGame controller and patiently waits for a position
update. The rendering itself also changed: Instead of FEN string parsing the
MBoardView uses the MPiece stored in the MPosition directly.

* src/game.[h|cc], src/position.[h|cc]: Moved signals and slots from MPosition
to MGame, our controller. This makes MPosition a passive component again,
suited to be passed around much like a container. Also added a test slot
(blackRookTest) so that we test our new MLogicAnalyzer.

* src/logic_analyzer.[h|cc]: This new class performs move validation. It knows
about MPosition and MPiece so it has a full understanding of a static chess
position and can also check transitions from a current position to possible
next positions.

* src/pieces/piece.[h|cc], src/pieces/rook.[h|cc]: Cleanup. Also added some
debug info which hopefully goes away soon.

* src/miniature.[h|cc], src/miniature.ui: Rewired some signals. Also activated
the blackRookTest from the menu.