Experimental feature: let MPosition update the scene graph
* src/pieces/piece.[h|cc], src/pieces/bishop.[h|cc]: Let the pieces derive from
QGraphicsSvgItem. This is what I had in mind in the very beginning, and it
allows faster board updates and better logic checks. It will also kill a lot
of code, hopefully.

* src/game.cc, src/position.[h|cc], src/board_view.[h|cc]: Let MPosition add
new pieces to a common board item (a generic QGraphicsSvgItem again) so that
it can update the piece's position directly after moving.