Improved piece selection and move performance
* src/*.[h|cc|ui], src/pieces/*.[h|cc]: On the N900, piece selection and move
performance was rather sluggish. Therefore, MPieces are now QGraphicsSvgItems
which can draw themselves. Upon making a move, the MPiece instance moves
accordingly on the scene graph, too. This allows to forgoe complete redrawing
of all pieces after each minor update. But it also brings in many sync
problems (differences in logical position in MPosition and position in
QGraphicsScene) which I wanted to avoid with a complete redraw action.

Now that the logical and graphical part of MPiece have been merged, most of the
code (and interaction!) between MGame, MBoardView and MGraphicsBoardItem has
changed dramatically. The latter is no longer owned by the MBoardView but is
instead controlled directly by MGame, killing some ugly signal propagation and
generally cutting code. The changes to MPosition have been on the lowest level
so that the MLogicAnalyzer is almost not affected.

There is still a bug w.r.t. to captured pieces and replays (prev/next
navigation), but the rest should work again.