Cleanup of MMainWindow files
If the autotools feature of conditional includes is to be meant to be used
instead of ifdefs, then this example for MMainWindow should be less confusing.

It should now be clear that the header file of MMainWindow serves as an
interface for the different implementations (not to be confused with a C++
"interface" class). The common, platform-independent parts can be put in
src/main_window.cc.

The generic/MeeGo implementation of MMainWindow is empty, as we still focus on
Maemo 5.