New: Game::AbstractSide to represent a side in the game
Idea is to choose a representation this time that is suitable for all kind of
game modes. Imagined backends:
* human vs. human (local, aka "pub mode"),
* human - engine (local),
* human - human (remote, either FICS or Telepathy again).

The important aspect here is that the actual game mode should not affect game
logic as much as it did in original Miniature code base. Different game modes
should just be a matter of using different Game::AbstractSide implementations.