Implement run in {background, foreground} API for Game::AbstractSide
Gnuchess runs computations all the time. Especially when we have multiple game
instances (but only one that is really active), we don't want the other
instances to waste time on computations. Once a side runs in background, we
want it to stop heavy computations and free resources, but not necessarily to
stop all operations.