Unify Game::AbstractSide
Each side implementation now takes a (shared) parser instance in the c'tor.
This unifies the other backends with the FICS backend, but also allows me to
disconnect from the parser when for example running in background. It is better
to have this logic deeply integrated than letting the GameManager mess around
with that, I think.