Get rid of old commandFound architecture
Let's only use the dispatcher for sending messages around. Otherwise it'll be
too confusing, eventually.
Decided that a Game's side should be simpler and suitable as a model for the
frontend => decoupled backend logic. That also killed the justification for the
AbstractSide interface.
Instead of a game manager, there is now just a game registry. But this one
makes sure that dispatcher talks to the active game and that the frontend hooks
up with the sides from the active game, too.