New: Allow frontend and backend to send moves directly to each other
Once the registry has set the active game on the frontend, we don't need to
route the moves through it. Can be easier done if we just let frontend and
backend exchange moves directly. Note how only frontend gets to know about the
game instance, as it can be assumed that the backend keeps its own game
representation anyway.