Visual signal to see clearer whose side is moving.
After playing a bit I've realized that it's not so evident to see when
the opponent has moved and it's your turn to play. This is an
implementation of a little blue signal next to the time. The current
implementation is functionally as broken as the local timers, will be
fixed when we fix those.