Fix castling when moves come from FICS side
FICS sends castling as o-o or O-O, but I didn't parse castling at all before.
Instead, I just relied on the position being correct, without the need for the
last move to be correct. Now last move is analyzed for castling, too.
Similar problems might arise in the future when it comes to moves with strange
notations, such as promotions.