Handle multiline chat messages coming from FICS
FICS introduces the escape character after newline, which makes it impossible
to know whether a given message will be continued or not. Going to assume that
next FICS token must begin with backlash after chat message, else it is single
line. Note that the Message command is send for each line though.