Split Scenario's play into play and respond
Make it more explicit that Scenario::play plays the canned data to the engine
until it encounters a WAIT_FOR_INPUT token. You then have to answer with the
right response, through Scenario::respond. Only if it matches the expected
response, you will be able to continue the scenario by calling Scenario::play
again.

If the response is wrong, the current scenario will be considered to have
failed. Scenario::finished will return true if a scenario has either failed or
was completed, which makes it suitable as the loop invariant, whereas
Scenario::result tells you if the scenario passed.

So Scenario::finished == true and Scenario::result == Passed means a scenario
was successful.

Also added this to Scenario class documentation.