Cleanup Game::AbstractEngine interface and fix documentation
Use readData (was: processToken) and writeData (new) to indicate
Miniature-engine communication. The latter proved to be useful in tests (that
is, being able to intercept messages Miniature eventually *would* send to the
engine), so I decided to make it a general part of the API, especially as it
represents the formerly missing Miniature-engine direction so nicely.

That being said, engines could still be read-only or write-only by choosing to
ignore read or write requests.

Also fixed class documentation to indicate our assumptions about Miniature
chess engines, namely that they usually come with a text stream interface.
This is true for Gnuchess and FICS at least.

Each method now states the communication direction (to/from engine).