Make Dispatcher requirement for engines explicit
Both CommandLine and FICS engine use the Dispatcher to send commands from the
engine to other Miniature components, so I think it only makes sense to have
setDispatcher in the AbstractEngine API.

Nice side effect: If no dispatcher is set, we can assume an engine is disabled.
This allowed me to remove setEnabled API.

Of course, with the dispatcher parameter being removed from the c'tor, engine
instance are now (nearly) trivial to construct.